\name{PoiClaClu-internal}
\title{Internal PoiClaClu functions}
\alias{NullModel}
\alias{NullModelTest}
\alias{print.poicla}
\alias{print.poiclacv}
\alias{print.poidist}
\description{Internal PoiClaClu functions}
\usage{
NullModel(x,type=c("mle","deseq","quantile"))
NullModelTest(null.out,x,xte,type=c("mle","deseq","quantile"))
}
\author{Daniela Witten}
\keyword{internal}



